requirements.txt
aioredis==2.0.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
Arabic-Stopwords==0.4.3
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Automat==22.10.0
autopep8==2.3.1
beautifulsoup4==4.12.3
billiard==4.2.0
blis==0.7.11
cairocffi==1.7.1
CairoSVG==2.7.1
catalogue==2.0.10
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.18.1
confection==0.1.5
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.8
cssselect==1.2.0
cssselect2==0.7.0
cymem==2.0.8
defusedxml==0.7.1
Django==4.2.7
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-filter==24.2
django-prometheus==2.3.1
django-redis==5.4.0
django-rest-knox==4.2.0
django-silk==5.1.0
django-timezone-field==7.0
djangorestframework==3.15.2
dnspython==2.6.1
drf-excel==2.4.1
drf-yasg==1.21.7
email_validator==2.2.0
et-xmlfile==1.1.0
fake-useragent==1.5.1
Faker==26.0.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
gprof2dot==2024.6.6
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanize==4.10.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
iso639-lang==2.2.3
itemadapter==0.9.0
itemloaders==1.3.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kombu==5.3.7
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
lxml==5.2.2
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
mdurl==0.1.2
meta-tags-parser==1.3.0
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.4
openpyxl==3.1.5
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parsel==1.9.1
pillow==10.4.0
preshed==3.0.9
prometheus_client==0.20.0
prompt_toolkit==3.0.47
Protego==0.3.1
psycopg2==2.9.9
PyArabic==0.6.15
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyDispatcher==2.0.7
Pygments==2.18.0
pygraphviz==1.13
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyquery==2.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
queuelib==1.7.0
readtime==3.0.0
redis==5.0.7
regex==2024.5.15
requests==2.32.3
requests-file==2.1.0
rich==13.7.1
scikit-learn==1.5.1
scipy==1.14.0
Scrapy==2.11.2
scrapy-fake-useragent==1.4.4
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.0
srsly==2.4.8
starlette==0.37.2
tblib==3.0.0
thinc==8.2.5
threadpoolctl==3.5.0
tinycss2==1.3.0
tldextract==5.1.2
tqdm==4.66.4
Twisted==24.3.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
validators==0.31.0
vine==5.1.0
w3lib==2.2.1
wasabi==1.1.3
watchfiles==0.22.0
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websockets==12.0
wrapt==1.16.0
zope.interface==6.4.post2
setuptools>=65.5.1 # not directly required, pinned by Snyk to avoid a vulnerability
zipp>=3.19.1 # not directly required, pinned by Snyk to avoid a vulnerability
